Changed enum Win32Error to a static class with constants and added Windows error code 1060: ERROR_SERVICE_DOES_NOT_EXIST.

+++ src/Server/Starcounter.Server/Windows/WindowsServiceHelper.cs
@@ -130,7 +130,7 @@ namespace Starcounter.Server.Windows {
                     );
                 if (br == false) {
                     lastError = Marshal.GetLastWin32Error();
-                    if (lastError != (int)Win32Error.ERROR_INSUFFICIENT_BUFFER)
+                    if (lastError != Win32Error.ERROR_INSUFFICIENT_BUFFER)
                         throw new Win32Exception(lastError);
 
                     Marshal.FreeHGlobal(buffer);
@@ -179,9 +179,8 @@ namespace Starcounter.Server.Windows {
 
                 if (br == false) {
                     lastError = Marshal.GetLastWin32Error();
-                    if (lastError != (int)Win32Error.ERROR_INSUFFICIENT_BUFFER)
-                        throw
-                            new Win32Exception(lastError);
+                    if (lastError != Win32Error.ERROR_INSUFFICIENT_BUFFER)
+                        throw new Win32Exception(lastError);
 
                     Marshal.FreeHGlobal(buffer);
                     bufferSize = bytesNeeded;
